github_runner_matrix: split into separate files
